feat(datavalueschart): finalize extracting data values chart logic
