--- src/components/Charts/properties.js
@@ -1,90 +1,0 @@
-import i18n from '@dhis2/d2-i18n'
-
-export const date = { key: 'date', label: i18n.t('Date') }
-export const mapViews = { key: 'mapViews', label: i18n.t('Map') }
-export const chartViews = { key: 'chartViews', label: i18n.t('Chart') }
-export const pivotTableViews = {
-    key: 'pivotTableViews',
-    label: i18n.t('Pivot Table'),
-}
-export const eventReportViews = {
-    key: 'eventReportViews',
-    label: i18n.t('Event Report'),
-}
-export const eventChartViews = {
-    key: 'eventChartViews',
-    label: i18n.t('Event Chart'),
-}
-export const dashboardViews = {
-    key: 'dashboardViews',
-    label: i18n.t('Dashboard'),
-}
-export const dataSetReportViews = {
-    key: 'dataSetReportViews',
-    label: i18n.t('Data Set Report'),
-}
-export const totalViews = { key: 'totalViews', label: i18n.t('Total') }
-export const averageMapViews = {
-    key: 'averageMapViews',
-    label: i18n.t('Average Map'),
-}
-export const averageChartViews = {
-    key: 'averageChartViews',
-    label: i18n.t('Average Chart'),
-}
-export const averagePivotTableViews = {
-    key: 'averagePivotTableViews',
-    label: i18n.t('Average Pivot Table'),
-}
-export const averageEventReportViews = {
-    key: 'averageEventReportViews',
-    label: i18n.t('Average Event Report'),
-}
-export const averageEventChartViews = {
-    key: 'averageEventChartViews',
-    label: i18n.t('Average Event Chart'),
-}
-export const averageDashboardViews = {
-    key: 'averageDashboardViews',
-    label: i18n.t('Average Dashboard'),
-}
-export const averageViews = {
-    key: 'averageViews',
-    label: i18n.t('Average'),
-}
-export const savedMaps = { key: 'savedMaps', label: i18n.t('Maps') }
-export const savedCharts = { key: 'savedCharts', label: i18n.t('Charts') }
-export const savedPivotTables = {
-    key: 'savedPivotTables',
-    label: i18n.t('Pivot Tables'),
-}
-export const savedEventReports = {
-    key: 'savedEventReports',
-    label: i18n.t('Event Reports'),
-}
-export const savedEventCharts = {
-    key: 'savedEventCharts',
-    label: i18n.t('Event Charts'),
-}
-export const savedDashboards = {
-    key: 'savedDashboards',
-    label: i18n.t('Dashboards'),
-}
-export const savedIndicators = {
-    key: 'savedIndicators',
-    label: i18n.t('Indicators'),
-}
-export const activeUsers = {
-    key: 'activeUsers',
-    label: i18n.t('Active Users'),
-}
-export const users = { key: 'users', label: i18n.t('Total Users') }
-export const position = { key: 'position', label: i18n.t('Position') }
-export const name = { key: 'name', label: i18n.t('Name') }
-export const views = { key: 'views', label: i18n.t('Views') }
-export const id = { key: 'id', label: i18n.t('ID') }
-export const created = { key: 'created', label: i18n.t('Created') }
-export const savedDataValues = {
-    key: 'savedDataValues',
-    label: i18n.t('Data Values'),
-}
